profile-service/requirements.txt
fastapi
uvicorn[standard]
sdv
lightgbm
joblib
pandas
numpy
scikit-learn
python-dotenv
